extract block production to its own file
